reduced platform specific stack trace code
