+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/Exceptions.kt
@@ -1,0 +1,24 @@
+package io.kotest.assertions
+
+/**
+ * Use this object to create exceptions on a target platform.
+ * This will create the most appropriate exception type, such as org.opentest4j.AssertionFailedError on
+ * platforms that support it, and defaulting to the basic kotlin AssertionError in the degenerative case.
+ */
+expect object Exceptions {
+
+   /**
+    * Creates an [AssertionError] from the given message. If the platform supports nested exceptions, the cause
+    * is set to the given [cause].
+    */
+   fun createAssertionError(message: String, cause: Throwable?): AssertionError
+
+   /**
+    * Creates the best error type supported on the platform (eg opentest4j.AssertionFailedException) from the
+    * given message and expected and actual values. If the platform supports nested exceptions, the cause
+    * is set to the given [cause].
+    *
+    * If the platform has jUnit4 or jUnit5 on the classpath, it will use exceptions from those platforms.
+    */
+   fun createAssertionError(message: String, cause: Throwable?, expected: Expected, actual: Actual): Throwable
+}
