+++ kotest-assertions/kotest-assertions-shared/src/nativeMain/kotlin/io/kotest/assertions/Exceptions.kt
@@ -1,0 +1,7 @@
+package io.kotest.assertions
+
+actual object Exceptions {
+   actual fun createAssertionError(message: String, cause: Throwable?): AssertionError = AssertionError(message)
+   actual fun createAssertionError(message: String, cause: Throwable?, expected: Expected, actual: Actual): Throwable =
+      createAssertionError(message, cause)
+}
