+++ kotest-mpp/src/jvmMain/kotlin/io/kotest/mpp/UserStackTraceConverter.kt
@@ -1,4 +1,4 @@
-package io.kotest.assertions
+package io.kotest.mpp
 
 object UserStackTraceConverter {
