+++ kotest-mpp/src/jvmMain/kotlin/io/kotest/mpp/stacktraces.kt
@@ -2,12 +2,17 @@
 
 package io.kotest.mpp
 
-/**
- * Returns the first line of a stack trace that isn't from the io.kotest packages.
- * On some platforms the stack trace may not be available.
- */
-actual fun Throwable.throwableLocation(): String? {
-   return (cause ?: this).stackTrace?.firstOrNull {
-      !it.className.startsWith("io.kotest")
-   }?.toString()
+actual object StackTraces {
+   actual fun Throwable.throwableLocation(): String? {
+      return (cause ?: this).stackTrace?.firstOrNull {
+         !it.className.startsWith("io.kotest")
+      }?.toString()
+   }
+
+   actual fun <T : Throwable> cleanStackTrace(throwable: T): T {
+      if (shouldRemoveKotestElementsFromStacktrace) {
+         throwable.stackTrace = UserStackTraceConverter.getUserStacktrace(throwable.stackTrace)
+      }
+      return throwable
+   }
 }
