+++ kotest-mpp/src/nativeMain/kotlin/io/kotest/mpp/reflection.kt
@@ -10,7 +10,7 @@ actual fun KClass<*>.fqn(): String? = this.qualifiedName
 /**
  * Returns the annotations on this class or empty list if not supported
  */
-actual fun KClass<*>.annotations(): List<Annotation> = this.annotations()
+actual fun KClass<*>.annotations(): List<Annotation> = emptyList()
 
 actual val <T : Any> KClass<T>.isDataClass: Boolean?
    get() = null
